+++ trunk/scintilla/src/Editor.cxx
@@ -5642,15 +5642,15 @@ void Editor::AddStyledText(char *buffer, int appendLength) {
 		for (i=0;i<textLength;i++) {
 			text[i] = buffer[i*2];
 		}
-		pdoc->InsertString(CurrentPosition(), text, textLength);
+		pdoc->InsertString(CurrentPosition(), text, (int)textLength);
 		for (i=0;i<textLength;i++) {
 			text[i] = buffer[i*2+1];
 		}
 		pdoc->StartStyling(CurrentPosition(), static_cast<char>(0xff));
-		pdoc->SetStyles(textLength, text);
+		pdoc->SetStyles((int)textLength, text);
 		delete []text;
 	}
-	SetEmptySelection(currentPos + textLength);
+	SetEmptySelection(currentPos + (int)textLength);
 }
 
 static bool ValidMargin(unsigned long wParam) {
